requirements.txt
numpy<2
pandas
scikit-learn
flask
kafka-python
requests
scikit-surprise
scikit-learn